Conflicts:
	src/main/java/org/zerock/domain/ReplyVo.java
	target/classes/org/zerock/domain/ReplyVo.class

+++ src/main/java/org/zerock/domain/UploadFileVo.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class UploadFileVo {
 	private int FILE_NO;
-	private int BOARD_NO;
+	private Long BOARD_NO;
 	private String FILE_NAME;
 }
